+++ src/bp/core/collaborators/index.ts
@@ -1,6 +1,3 @@
 export * from './repositories/strategy_users'
 export * from './repositories/workspace_users'
 export * from './repositories/workspace_invite_codes'
-export * from './tables/strategy_users'
-export * from './tables/workspace_users'
-export * from './tables/workspace_invite_codes'
