+++ src/bp/core/events/event-collector.ts
@@ -12,41 +12,6 @@ import { EventRepository } from './event-repository'
 
 type BatchEvent = sdk.IO.StoredEvent & { retry?: number }
 
-export const StepScopes = {
-  Received: 'received',
-  StateLoaded: 'stateLoaded',
-  Middleware: 'mw',
-  Dialog: 'dialog',
-  Action: 'action',
-  Hook: 'hook',
-  EndProcessing: 'completed'
-}
-
-export const StepStatus = {
-  Started: 'start',
-  Completed: 'completed',
-  Error: 'error',
-  TimedOut: 'timedOut',
-  Swallowed: 'swallowed',
-  Skipped: 'skipped'
-}
-
-export const addStepToEvent = (event: sdk.IO.Event, scope: string, name?: string, status?: string) => {
-  if (!event?.debugger) {
-    return
-  }
-
-  event.processing = {
-    ...(event.processing || {}),
-    [`${scope}${name ? `:${name}` : ''}${status ? `:${status}` : ''}`]: {
-      ...(event?.activeProcessing || {}),
-      date: new Date()
-    }
-  }
-
-  event.activeProcessing = {}
-}
-
 export const addLogToEvent = (logEntry: string, event: sdk.IO.Event) => {
   if (event?.debugger && event?.activeProcessing) {
     event.activeProcessing.logs = [...(event.activeProcessing.logs ?? []), logEntry]
