+++ src/bp/core/events/index.ts
@@ -3,3 +3,4 @@ export * from './event-engine'
 export * from './event-repository'
 export * from './queue'
 export * from './queue/memory-queue'
+export * from './utils'
